add filtering to users section
