+++ packages/admin-panel/src/pages/resources/UsersPage.js
@@ -40,6 +40,10 @@ const FIELDS = [
     source: 'position',
   },
   {
+    Header: 'Platform',
+    source: 'primary_platform',
+  },
+  {
     Header: 'Employer',
     source: 'employer',
     type: 'tooltip',
@@ -135,7 +139,7 @@ const CREATE_CONFIG = {
   },
 };
 
-export const UsersPage = ({ getHeaderEl }) => (
+export const UsersPage = ({ getHeaderEl, ...props }) => (
   <ResourcePage
     title="Users"
     endpoint="users"
@@ -145,6 +149,7 @@ export const UsersPage = ({ getHeaderEl }) => (
     editConfig={EDIT_CONFIG}
     createConfig={CREATE_CONFIG}
     getHeaderEl={getHeaderEl}
+    {...props}
   />
 );
